requirements.txt
pandas>=1.3.0
spacy>=3.1.0
sentence-transformers>=2.2.0
pdfminer.six>=20200726
mammoth>=1.4.0